requirements.txt
Django==5.0.2
djangorestframework==3.14.0
python-dotenv==1.0.1
requests==2.31.0
django-cors-headers==4.3.1
celery==5.3.6
redis==5.0.1 